feat: File Select Panel 을 Control File Panel로 기능 병합
- Character Select Panel에서 뒤로가기 시 Control File Panel과 File Select Panel이 다른 Stack이기 때문에 분리되어 의도상 맞지 않아 병합
