--- Assets/Min/Scripts/UIs/PanelBehaviours/FileSelectPanelButtons.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a09a1fbf968e59043acceba394797265
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
